Break shims down into separate modules
Most are moved to own files in `/shims/`
